[regression-test](fix) fix case, pick #46885

+++ regression-test/suites/alter_p2/test_alter_table_property.groovy
@@ -16,6 +16,10 @@
 // under the License.
 
 suite ("test_alter_table_property") {
+    if (isCloudMode()) {
+        return
+    }
+
     String tableName = "test_alter_table_property_table"
     sql "DROP TABLE IF EXISTS ${tableName}"
     sql """
@@ -46,11 +50,21 @@ suite ("test_alter_table_property") {
         }
         sum
     }
+    def replication_num = 1
+    def forceReplicaNum = getFeConfig('force_olap_table_replication_num').toInteger()
+    if (forceReplicaNum > 0) {
+        replication_num = forceReplicaNum
+    }
 
-    assertEquals(1, queryReplicaCount("p1"))
+    assertEquals(replication_num, queryReplicaCount("p1"))
 
     sql """ ALTER TABLE ${tableName} ADD PARTITION p2 VALUES LESS THAN ("200") """
-    assertEquals(1, queryReplicaCount("p2"))
+    assertEquals(replication_num, queryReplicaCount("p2"))
+
+    def res = sql """show backends;"""
+    if (res.size() < 3) {
+        return
+    }
 
     sql """ ALTER TABLE ${tableName} SET ( "default.replication_allocation" = "tag.location.default: 2" ) """
     sql """ ALTER TABLE ${tableName} ADD PARTITION p3 VALUES LESS THAN ("300") """
@@ -63,7 +77,7 @@ suite ("test_alter_table_property") {
         }
     }
     assertEquals(2, queryReplicaCount("p1"))
-    assertEquals(1, queryReplicaCount("p2"))
+    assertEquals(replication_num, queryReplicaCount("p2"))
 
     sql "DROP TABLE ${tableName}"
 }
